Update Filler CT [var inc]
Meant to be a Client Script, this code will look at a target variable and only reference the subset of selected information within that column. 

Example: 
Assume variable fields A and B
Field A holds manufacturer, in which the customer selects "Toshiba"
Field B (a list collector) will now only show options which fall into the "Toshiba" category in the selected table.